+++ src/torchpme/potentials/coulomb.py
@@ -103,5 +103,56 @@ class CoulombPotential(Potential):
             )
         return torch.pi * self.smearing**2
 
+    @staticmethod
+    def pbc_correction(
+        periodic: Optional[torch.Tensor],
+        positions: torch.Tensor,
+        cell: torch.Tensor,
+        charges: torch.Tensor,
+    ) -> torch.Tensor:
+        # "2D periodicity" correction for 1/r potential
+        if periodic is None:
+            periodic = torch.tensor([True, True, True], device=cell.device)
+
+        n_periodic = torch.sum(periodic).item()
+        if n_periodic == 3:
+            periodicity = 3
+            nonperiodic_axis = None
+        elif n_periodic == 2:
+            periodicity = 2
+            nonperiodic_axis = torch.where(~periodic)[0]
+            max_distance = torch.max(positions[:, nonperiodic_axis]) - torch.min(
+                positions[:, nonperiodic_axis]
+            )
+            cell_size = torch.linalg.norm(cell[nonperiodic_axis])
+            if max_distance > cell_size / 3:
+                raise ValueError(
+                    f"Maximum distance along non-periodic axis ({max_distance}) "
+                    f"exceeds one third of cell size ({cell_size})."
+                )
+        else:
+            raise ValueError(
+                "K-space summation is not implemented for 1D or non-periodic systems."
+            )
+
+        if periodicity == 2:
+            charge_tot = torch.sum(charges, dim=0)
+            axis = nonperiodic_axis
+            z_i = positions[:, axis].view(-1, 1)
+            basis_len = torch.linalg.norm(cell[axis])
+            M_axis = torch.sum(charges * z_i, dim=0)
+            M_axis_sq = torch.sum(charges * z_i**2, dim=0)
+            V = torch.abs(torch.linalg.det(cell))
+            E_slab = (4.0 * torch.pi / V) * (
+                z_i * M_axis
+                - 0.5 * (M_axis_sq + charge_tot * z_i**2)
+                - charge_tot / 12.0 * basis_len**2
+            )
+        else:
+            E_slab = torch.zeros_like(charges)
+
+        return E_slab
+
     self_contribution.__doc__ = Potential.self_contribution.__doc__
     background_correction.__doc__ = Potential.background_correction.__doc__
+    pbc_correction.__doc__ = Potential.pbc_correction.__doc__