+++ docs/src/references/changelog.rst
@@ -47,6 +47,11 @@ Fixed
 * Update metatensor bindings to metatomic
 
 
+Fixed
+#####
+
+* Fix exclusion_radius to work in a direct calculator
+
 `Version 0.3.0 <https://github.com/lab-cosmo/torch-pme/releases/tag/v0.3.0>`_ - 2025-02-21
 ------------------------------------------------------------------------------------------